Move the remaining package checking machinery to its own module.
