+++ src/allmydata/test/test_runner.py
@@ -23,6 +23,7 @@ from allmydata.util import fileutil, pollmixin
 from allmydata.util.encodingutil import unicode_to_argv, unicode_to_output, \
     get_filesystem_encoding
 from allmydata.test import common_util
+from allmydata.version_checks import normalized_version
 import allmydata
 from allmydata import __appname__
 from .common_util import parse_cli, run_cli
@@ -112,8 +113,6 @@ class BinTahoe(common_util.SignalMixin, unittest.TestCase, RunBinTahoeMixin):
     def test_path(self):
         d = self.run_bintahoe(["--version-and-path"])
         def _cb(res):
-            from allmydata import normalized_version
-
             out, err, rc_or_sig = res
             self.failUnlessEqual(rc_or_sig, 0, str(res))
